Rename 'R' directories and package to 'rlang'
