+++ jplag.frontend.rlang/src/main/java/de/jplag/rlang/RTokenConstants.java
@@ -1,4 +1,4 @@
-package de.jplag.R;
+package de.jplag.rlang;
 
 import de.jplag.TokenConstants;
